removed try-catch from constructor, used better validation

+++ BloomFilter/HashFunctions/Djb2.cs
@@ -2,9 +2,9 @@ namespace BloomFilter.HashFunctions;
 
 public static class Djb2
 {
-    public static ulong Hash(string input)
+    public static uint Hash(string input)
     {
-        ulong hash = 5381; // prime
+        uint hash = 5381; // prime
         foreach (char c in input)
         {
             hash = (hash << 5) + hash + c;
